Refactor import statements in page.tsx and layout.tsx
