+++ app/[controller]/page.tsx
@@ -1,4 +1,4 @@
-import { loadApiSchema } from "@/lib/utils";
+import { loadApiSchema } from "@/lib/utils/load";
 
 export default async function Page({ params }: { params: Promise<{ controller: string }> }) {
   const { controller } = await params;
