+++ app/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import { IconDescriptor } from "next/dist/lib/metadata/types/metadata-types";
 import { EndpointsDataProvider } from "@/contexts";
 import { Sidebar } from "@/components";
-import { loadApiSchema, loadProjectData } from "@/lib/utils";
+import { loadApiSchema, loadProjectData } from "@/lib/utils/load";
 import "./globals.css";
 
 export interface CustomIconDescriptorType extends IconDescriptor {
