+++ lib/utils/index.ts
@@ -1,1 +1,1 @@
-export * from "./load";
+export * from "./typo";
